moved file related code into Files.mjs from DOM.mjs
