+++ src/Files.mjs
@@ -106,3 +106,37 @@ export const exists = (path) =>
 export const rm = (path) => fsp.rm(path, { force: true, recursive: true }).then(
   onfulfilled => path,
   onrejected => onrejected);
+
+// TODO exporting only for unit testing
+export const removeParameters = (url) => {
+  const base = path.basename(url);
+  const i = base.indexOf('?');
+  return i === -1 ? url :
+    path.join(path.dirname(url), base.substring(0, i));
+}
+/**
+ * Extracts relative paths from tagName[attrName] elements
+ * under the given dom node.
+ *
+ * @param {Cheerio} dom The Cheerio instance of DOM.
+ */
+const getLocalFiles = (dom) => {
+  const localFiles = [];
+  dom.find(`link[href], script[src], img[src]`).each((i, ele) => {
+    const node = new Cheerio(ele);
+    const url = node.attr('href') || node.attr('src');
+    if (!url.match(notRelative) && !path.isAbsolute(url)) {
+      localFiles.push(removeParameters(url));
+    }
+  });
+  return localFiles;
+};
+
+export const copyRelativeFiles = (basefile, outDir) => (dom) => {
+  const toAbsoluteInOutDir = (relativeFile) => path.join(outDir, relativeFile);
+  const toAbsoluteInSrcDir = relative2absolute(basefile);
+
+  getLocalFiles(dom).forEach(file =>
+    copyIfNewer(toAbsoluteInSrcDir(file))
+    (toAbsoluteInOutDir(file)).catch(e => console.log(`    Local file linked from the document is missing: ${toAbsoluteInSrcDir(file)}`)));
+};
