+++ test/testDOM.mjs
@@ -18,13 +18,12 @@ import {
   makeDocument,
   getFirstContentId,
   getContentNode$,
-  removeParameters,
 } from '../src/DOM.mjs';
 import { append$ } from '../src/DomFunc.mjs';
 import { pipe } from '../src/FP.mjs';
 import cheerio from 'cheerio';
 import { Cheerio } from '../node_modules/cheerio/lib/cheerio.js';
-import { rm, exists } from '../src/Files.mjs';
+import { rm, exists, removeParameters } from '../src/Files.mjs';
 import { mkdirs } from '../src/Files.mjs';
 import {
   getFootnoteDefIds,
